# Optimize magnet generation